pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "FlowDetection"
version = "0.1.0"
description = "CNN-based flow detection from camera images"
requires-python = ">=3.12,<3.13"
dependencies = [
	"matplotlib>=3.8",
	"numpy>=1.26",
	"PyYAML>=6.0",
	"tensorflow==2.16.1",
]

[project.optional-dependencies]
scripts = [
	"openpyxl>=3.1",
	"pandas>=2.2",
]
experimental = [
	"scikeras==0.13.0",
	"scikit-learn==1.5.2",
	"scipy==1.14.1",
]

[tool.setuptools.packages.find]
# Tell setuptools to look in src/ for packages, not the repo root.
# This means `import FlowDetection` resolves to src/FlowDetection/
where = ["src"]
